requirements.txt
fastapi
uvicorn

deta
python-multipart

passlib
pydantic
python-jose

folium
pandas

requests

python-dotenv
